コミット

+++ database/migrations/2024_12_17_223257_create_orders_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('method', 50);
             $table->string('address', 1000);
             $table->string('total_products', 1000);
-            $table->decimal('total_price');
+            $table->integer('total_price');
             $table->date('placed_on');
             $table->string('status')->default('pending');
